// Capstone Project/src/app/pages/admin/admin-room-category-management-page/screen/types.ts
export enum adminPageRoomCategoryEnum {
  GET_DATA_ROOM_CATEGORIES_REQUEST = 'APP/ADMIN_PAGE/GET_DATA_ROOM_CATEGORIES_REQUEST',
  GET_DATA_ROOM_CATEGORIES_SUCCESS = 'APP/ADMIN_PAGE/GET_DATA_ROOM_CATEGORIES_SUCCESS',
  GET_DATA_ROOM_CATEGORIES_FAILURE = 'APP/ADMIN_PAGE/GET_DATA_ROOM_CATEGORIES_FAILURE',

  CREATE_DATA_ROOM_CATEGORIES_REQUEST = 'APP/ADMIN_PAGE/CREATE_DATA_ROOM_CATEGORIES_REQUEST',
  CREATE_DATA_ROOM_CATEGORIES_SUCCESS = 'APP/ADMIN_PAGE/CREATE_DATA_ROOM_CATEGORIES_SUCCESS',
  CREATE_DATA_ROOM_CATEGORIES_FAILURE = 'APP/ADMIN_PAGE/CREATE_DATA_ROOM_CATEGORIES_FAILURE',

  UPDATE_DATA_ROOM_CATEGORIES_REQUEST = 'APP/ADMIN_PAGE/UPDATE_DATA_ROOM_CATEGORIES_REQUEST',
  UPDATE_DATA_ROOM_CATEGORIES_SUCCESS = 'APP/ADMIN_PAGE/UPDATE_DATA_ROOM_CATEGORIES_SUCCESS',
  UPDATE_DATA_ROOM_CATEGORIES_FAILURE = 'APP/ADMIN_PAGE/UPDATE_DATA_ROOM_CATEGORIES_FAILURE',

  DELETE_DATA_ROOM_CATEGORIES_REQUEST = 'APP/ADMIN_PAGE/DELETE_DATA_ROOM_CATEGORIES_REQUEST',
  DELETE_DATA_ROOM_CATEGORIES_SUCCESS = 'APP/ADMIN_PAGE/DELETE_DATA_ROOM_CATEGORIES_SUCCESS',
  DELETE_DATA_ROOM_CATEGORIES_FAILURE = 'APP/ADMIN_PAGE/DELETE_DATA_ROOM_CATEGORIES_FAILURE',

  CLEAR_ROOM_CATEGORY_STATE = 'APP/ADMIN_PAGE/CLEAR_ROOM_CATEGORY_STATE',
}